requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
black==21.10b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
filelock==3.3.2
identify==2.3.3
idna==3.3
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.15.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
